VitaScanAI/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.0.0
scikit-learn==1.5.2
xgboost==2.1.3
transformers==4.47.1
joblib==1.4.2
boto3==1.35.81
python-dotenv==1.0.1
httpx==0.28.1
numpy==1.26.4
pandas==2.2.3
aiofiles==24.1.0
PyJWT==2.10.1
torch==2.5.1 --index-url https://download.pytorch.org/whl/cpu
